requirements.txt
pytest
pytest-playwright
playwright
pytest-html
pytest-base-url
allure-pytest
